relocate object and refactor functions
