+++ src/scenes/Nivel1.js
@@ -1,106 +1,4 @@
-class Nave extends Phaser.GameObjects.Sprite {
-  constructor(scene, x, y, meteoroGroup) {
-    super(scene, x, y, "nave");
-
-    scene.add.existing(this);
-    scene.physics.world.enable(this);
-
-   // this.body.setGravityY(0);
-    this.body.setCollideWorldBounds(true);
-    this.body.setVelocityX(300);
-    this.setScale(0.45);
-
-    this.isShooting = false;
-    this.shootDelay = 300;
-    this.lastShootTime = 400;
-
-    this.meteoroGroup = meteoroGroup; // Asignar el grupo de meteoros
-  }
-  init() {
-    this.vidas = 3;
-  }
-  create() {
-    this.nave = this.physics.add.sprite(163.33333333, 510.666666666, "nave");
-    this.nave.body.allowGravity = false;
-    this.nave.setCollideWorldBounds(true);
-  }
-
-  update(time) {
-    const teclas = this.scene.input.keyboard.createCursorKeys();
-    const camera = this.scene.cameras.main;
-
-    if (teclas.up.isDown && this.y > camera.y) {
-      this.body.velocity.y = -400; // Velocidad hacia arriba
-    } else if (teclas.down.isDown && this.y < camera.y + camera.height) {
-      this.body.velocity.y = 400; // Velocidad hacia abajo
-    } else {
-      this.body.velocity.y = 0;
-    }
-
-    if (teclas.space.isDown) {
-      // Verificación del tiempo suficiente desde el último disparo
-      if (time > this.lastShootTime + this.shootDelay) {
-        this.shoot();
-        this.lastShootTime = time;
-      }
-    }
-  }
-
-  shoot() {
-    // Crear y configurar el disparo
-    const disparo = this.scene.physics.add.sprite(this.x + 15, this.y + 1, 'Disparo');
-    disparo.setVelocityX(1500); // Velocidad del disparo en el eje X
-    disparo.setScale(0.3);
-
-    // Destrucción el disparo después de cierto tiempo para evitar la acumulación de disparos en el mundo
-    this.scene.time.addEvent({
-      delay: 2000, // Tiempo en milisegundos antes de destruir el disparo
-      callback: () => {
-        disparo.destroy();
-      },
-    });
-  }
-  handleCollision(nave, meteoro) {
-    // Restar una vida
-    this.vidas--;
-  
-    // Ocultar y desactivar el meteoro colisionado
-    meteoro.setVisible(false);
-    meteoro.setActive(false);
-  
-    // Comprobar si aún quedan vidas
-    if (this.vidas > 0) {
-      // Hacer parpadear la nave
-      this.blinkNave();
-  
-      // Esperar 2 segundos antes de permitir otra colisión
-      this.scene.time.delayedCall(2000, () => {
-        // Hacer visible la nave nuevamente
-        nave.setActive(true);
-        nave.setVisible(true);
-      }, this);
-    }
-  }
-
-  blinkNave() {
-    const nave = this.nave;
-    nave.setVisible(false);
-
-    const blinkInterval = setInterval(() => {
-      nave.setVisible(!nave.visible);
-    }, 200);
-
-    this.scene.time.delayedCall(2000, () => {
-      clearInterval(blinkInterval);
-      nave.setVisible(true);
-    }, this);
-  }
-
-  gameOver() {
-    if (this.vidas <= 0) {
-    this.scene.start('nivel2');}
-  }
-}
+import Nave from "../objects/Nave.js";
 
 export default class Nivel1 extends Phaser.Scene {
   constructor() {
@@ -109,6 +7,7 @@ export default class Nivel1 extends Phaser.Scene {
   init() {
     this.playerSurvived = false;
     this.isNextLevelEnabled = false;
+    this.vidas = 3;
   }
   
   create() {
@@ -203,7 +102,7 @@ export default class Nivel1 extends Phaser.Scene {
       this
     );    
     // Colisión entre la nave y los meteoros del grupo
-    this.physics.add.overlap(this.nave, this.meteoroGroup, this.nave.handleCollision, null, this);
+    this.physics.add.overlap(this.nave, this.meteoroGroup, this.handleCollision, null, this);
   }
 
   
@@ -238,11 +137,40 @@ export default class Nivel1 extends Phaser.Scene {
     };
     
   }
+
   update(time) {
     this.nave.update(time);
     const cameraOffsetX = -500; // Desplazamiento horizontal desde la posición de la nave
     const cameraOffsetY = 0; // Desplazamiento vertical desde la posición de la nave
     this.cameras.main.setFollowOffset(cameraOffsetX, cameraOffsetY);
+  }
+
+  handleCollision(nave, meteoro) {
+    console.log('Colisión con meteoro')
+    // Restar una vida
+    this.vidas--;
+  
+    // Ocultar y desactivar el meteoro colisionado
+    meteoro.disableBody(true, true);
+  
+    // Comprobar si aún quedan vidas
+    if (this.vidas > 0) {
+      // Hacer parpadear la nave
+      this.nave.blinkNave();
+  
+      // Esperar 2 segundos antes de permitir otra colisión
+      this.time.delayedCall(2000, () => {
+        // Hacer visible la nave nuevamente
+        this.nave.setActive(true);
+        this.nave.setVisible(true);
+      }, this);
+    } else {
+      this.gameOver();
+    }
+  }
+
+  gameOver() {
+    this.scene.start('nivel2');
   }
   
 }
